+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Adds `danger-junit` to the `danger/danger` repo, requiring changes to the plugin testing systems - orta
 * Improves the "is a shared Dangerfile from the Danger Repo" check #366 - orta
+* Replaces `redcarpet` through `kramdown` to avoid jruby foo - LeFnord
 
 ## 2.1.0